use native json marshaling to map stubbed output
When an output field had a '_' in its name, it was not mapped
correctly by 'mapstructure' to the protoc generated class because,
in this class, the fields are translated in CamelCase.
Hower, the fields have json tag with their original name.
Json marshalling is then the way to fix this issue.