Support OpenTelemetry/Zipkin tracing in gripmock
Instrument gRPC requests with OpenTelemetry trace support.

Supporting OpenTelemetry required some changes in how
the server is built. Instead of a bare "go run", the
server must be built with "go mod init", "go mod tidy",
"go mod download", and "go build".

The following exporters are supported by default:

* OpenTelemetry collector (gRPC and HTTP)
  https://opentelemetry.io/docs/reference/specification/sdk-environment-variables/#otlp-exporter
* Zipkin
  https://opentelemetry.io/docs/reference/specification/sdk-environment-variables/#zipkin-exporter
* Stdout

While metrics exporter support is not introduced by this commit, it
would be trivial to add metrics for gRPC request processing etc
using the same OpenTelemetry API.

The instrumentation provided does not capture request headers and
payloads or response headers and payloads. Another gRPC interceptor
could be used to optionally record this information.

The stubs API server is not instrumented, as it is not considered
interesting for application tracing.