Update README section regarding new headers matching rule
A new headers matching rule was implemented recently, in #177.
However, looking at the implementation and at the [stub json schema](https://gripmock.org/schema/stub.json),
it seems the documentation is wrong. In the documentation it
states that "headers" is a field nested under "inputs", however it
is a field on the root object (a sibling of "inputs").

I can confirm that this is indeed the case because I've been
trying to use the "headers" matcher according to the docs, 
and it never gets registered. However, using the stub with the
"headers" at the root object, the stub gets registered.